% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behavior.R
\name{actogram_loadFromMetadata}
\alias{actogram_loadFromMetadata}
\title{Load behavior activity data from metadata}
\usage{
actogram_loadFromMetadata(path, data_dir = "./", ...)
}
\arguments{
\item{path}{Path to metadata, must be commonly used delimited or excel.}

\item{data_dir}{Path to the behavior data dir.}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
\code{behavr::behavr} behavior data table.
}
\description{
Load behavior activity data from metadata
}
\examples{
# TODO
}
