% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actogram.R
\name{actogram_plot}
\alias{actogram_plot}
\title{Plotting actogram}
\usage{
actogram_plot(
  behavr,
  file = NULL,
  file.options = list(device = tools::file_ext(file), height = 9.14, width = 10.6)
)
}
\arguments{
\item{behavr}{\code{behavr::behavr} behavior data table.}

\item{file}{Optional path to save the plot. Backend is \code{ggplot2::ggsave()}.}

\item{file.options}{List to specify options for saving the plot.}
}
\value{
ggplot plot object.
}
\description{
Plotting actogram
}
\examples{
# TODO
}
