% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behavior.R
\name{behavior_loadMetadata}
\alias{behavior_loadMetadata}
\title{Load behavior metadata}
\usage{
behavior_loadMetadata(path)
}
\arguments{
\item{path}{Path to metadata file. See body for column spec.}
}
\value{
Metadata tibble.
}
\description{
Load behavior metadata
}
\examples{
# Not exported.
}
