% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctcf.R
\name{ctcf_plot}
\alias{ctcf_plot}
\title{Standardized plotting of CTCF data}
\usage{
ctcf_plot(data, x, y, fill = NA, template = "boxPoint")
}
\arguments{
\item{data}{tibble of CTCF data}

\item{x}{x-axis variable for plotting}

\item{y}{y-axis variable for plotting}

\item{fill}{fill variable for plotting}

\item{template}{which template to use, see details.}
}
\value{
ggplot object
}
\description{
This is de facto the versatile X-Y plot generator of this package. Supports
multiple plot templates.
}
\details{
All possible templates are not exported functions \code{plot.} in this package.
Supported ones are:
\itemize{
\item \code{"linePoint"} - uses \code{plot.linePoint()} - point plot with summary line
\item \code{"boxPoint"} - uses \code{plot.boxPoint()} - point plot with box
}
}
\examples{
# Will warn about the `am` continuous color variable
ctcf_plot(mtcars, cyl, mpg, am, template = "linePoint")
# Will create a boxPoint plot. Note that for boxPoint X must be discrete.
ctcf_plot(
mtcars |> dplyr::mutate(gear = as.character(gear), am = as.character(am)),
am, mpg, gear)
}
