% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctcf.R
\name{ctcf_read_csv}
\alias{ctcf_read_csv}
\title{Read CTCF results table}
\usage{
ctcf_read_csv(path, ...)
}
\arguments{
\item{path}{Path to the CTCF analysis csv}

\item{...}{Forwarded to \code{readr::read_csv}}
}
\value{
tibble of CTCF table
}
\description{
Read CTCF results table
}
\examples{
#TODO
}
