% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{.parse_meta}
\alias{.parse_meta}
\title{Parse spot array JSON data to metaRcrd.}
\usage{
.parse_meta(ser.coord, ser.dist)
}
\arguments{
\item{ser.coord}{Serialized coord.}

\item{ser.dist}{Serialized dist.}
}
\value{
A \code{y3628::metaRcrd} object.
}
\description{
Parse spot array JSON data to metaRcrd.
}
\examples{
# Does not apply. Internal use only.
}
