% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.NCT.R
\name{.selectNCT.index}
\alias{.selectNCT.index}
\title{Get index for spotInRoi nuclear/full-cell data subsetting.}
\usage{
.selectNCT.index(dfNucCyt, what)
}
\arguments{
\item{dfNucCyt}{Parsed spotInRoi nuclear/full-cell data.}

\item{what}{What to subset. must be \code{nuclear}, \code{cytoplasm}, or \code{full}.}
}
\value{
Indices for subsetting.
}
\description{
Get index for spotInRoi nuclear/full-cell data subsetting.
}
\examples{
# Does not apply. Internal use only.
}
