% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnd.compute.R
\name{ij_nnd.compute}
\alias{ij_nnd.compute}
\title{Compute NND metrics for a pair of spot sets.}
\usage{
ij_nnd.compute(foreach, to, z.proximal)
}
\arguments{
\item{foreach}{For each spot in \code{foreach},}

\item{to}{compute its distance to the nearest neighbor in \code{to} spot set.}

\item{z.proximal}{Only spots in close enough z-proximity is computed.}
}
\value{
\code{y3628::metaRcrd} object storing the NND results. For spots that have
no Z-proximal spots in the \code{to} spot set, NAs are returned.
}
\description{
Compute NND metrics for a pair of spot sets.
}
\examples{
# TODO
}
