% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RoiPointsSaver.R
\name{ij_RoiPointsSaver}
\alias{ij_RoiPointsSaver}
\title{Initialize an ImageJ ROI point set saver function}
\usage{
ij_RoiPointsSaver(use_python)
}
\arguments{
\item{use_python}{Path to the Python binary to use.}
}
\value{
A ROI saver function.
}
\description{
Initialize an ImageJ ROI point set saver function
}
\examples{
# TODO
}
