% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhythm_JTK.R
\name{jtk.init}
\alias{jtk.init}
\title{JTK Cycle V3.1 Adaptation \code{jtk.init}}
\usage{
jtk.init(periods, interval = 1)
}
\arguments{
\item{periods}{Periods to test on (in units of spacing \code{interval}).}

\item{interval}{Spacing of time points (in physical unit like hours).

For example, \code{jtk.init(9:15, 2)} means that data time point spacing is
2-hour and possible periods would be 18-hour to 30-hour (2-hour spacing).}
}
\value{
None; Run only for side effects on \code{JTK_env}.
}
\description{
Initialize the JTK environment for all periods
}
\examples{
# Refer to the JTK Cycle guide.
}
