% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhythm_JTK.R
\name{jtkdist}
\alias{jtkdist}
\title{JTK Cycle V3.1 Adaptation \code{jtkdist}}
\usage{
jtkdist(timepoints, reps = 1, normal = FALSE, alt = FALSE)
}
\arguments{
\item{timepoints}{Number of time points in the data.}

\item{reps}{Number of replicates for each time point.}

\item{normal}{Not used. Refer to JTK Cycle paper.}

\item{alt}{Not used. Refer to JTK Cycle paper.}
}
\value{
Not used. Refer to JTK Cycle paper.
}
\description{
JTK Cycle V3.1 Adaptation \code{jtkdist}
}
\examples{
# Refer to the JTK Cycle guide.
}
