% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhythm_JTK.R
\name{jtkget}
\alias{jtkget}
\title{JTK Cycle V3.1 Adaptation}
\usage{
jtkget(get.AMP.CI = FALSE)
}
\arguments{
\item{get.AMP.CI}{Bool, whether to get amplitude confidence interval,
which will be length=2 numeric vector.}
}
\value{
Named numeric vector of current \code{jtkx} results stored in \code{JTK_env}.
If \code{get.AMP.CI==TRUE}, only return the amplitude confidence interval.
}
\description{
This is getter function to fetch computation results of \code{jtkx}.
}
\details{
In this adaptation, \code{jtkx} results are stored internally in a package
environment \code{JTK_env}. This function is a getter of the enrivonment.
}
\examples{
# Refer to the JTK Cycle guide.
}
