% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhythm_JTK.R
\name{jtkstat}
\alias{jtkstat}
\title{JTK Cycle V3.1 Adaptation \code{jtkstat}}
\usage{
jtkstat(z)
}
\arguments{
\item{z}{Numeric vector of data at each time point.}
}
\value{
None; Run only for side effects on \code{JTK_env}.
}
\description{
jtkstat: calculate the p-values for all (period,phase) combos.
v3.1 modified to analyze data with missing values.
}
\examples{
# Internal use by `jtkx` only.
}
