% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhythm_JTK.R
\name{jtkx}
\alias{jtkx}
\title{JTK Cycle V3.1 Adaptation \code{jtkx}}
\usage{
jtkx(z, ampci = FALSE, conf = 0.8)
}
\arguments{
\item{z}{Numeric vector of data at each time point.}

\item{ampci}{Boolean, whether to compute amplitude CI and p-value.}

\item{conf}{Confidence level of amplitude. Only useful if \code{ampci == TRUE}.}
}
\value{
None; Run only for side effects on \code{JTK_env}.
}
\description{
jtkx: integration of jtkstat and jtkdist for repeated use.
}
\examples{
# Refer to the JTK Cycle guide.
}
