% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{nlsGroup}
\alias{nlsGroup}
\title{\code{nls()} fitting of multiple groups}
\usage{
nlsGroup(.data, var.group, formula, ...)
}
\arguments{
\item{.data}{data.frame with variables to fit with additional group variables.}

\item{var.group}{group variables. Follow \code{y3628::grouper()}.}

\item{formula}{Formula to run \code{nls()}}

\item{...}{Additional parameters forwarded to \code{nls()}. Most notably, you
need to provide start values unless you are using self-starting models.
See \code{stats::nls()}.}
}
\value{
Data frame of fitting parameters, with group variables.
}
\description{
\code{nls()} fitting of multiple groups
}
\examples{
# Puromycin, self-starting SSmicmen kinetics
nlsGroup(Puromycin, state, rate~SSmicmen(conc,Vm,K))
# Custom formula. Must provide start values.
michaelis_menten <- rate ~ (Vm)*conc/(K+conc)
nlsGroup(Puromycin, state, michaelis_menten, start=list(Vm=100,K=0.01))
}
