% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnd.export.R
\name{nnd.order}
\alias{nnd.order}
\title{Order permutation for a NND result record}
\usage{
nnd.order(nnd.results, ...)
}
\arguments{
\item{nnd.results}{\code{y3628::metaRcrd} of NND result.}

\item{...}{Passed to \code{base::order}}
}
\value{
Order permutation using the \code{nnd} field.
}
\description{
Order permutation for a NND result record
}
\examples{
# Does not apply. Internal use only.
}
