% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behavior.R
\name{periodogram_loadClocklab}
\alias{periodogram_loadClocklab}
\title{Load Clocklab periodogram batch export table}
\usage{
periodogram_loadClocklab(data, meta, ...)
}
\arguments{
\item{data}{Path to the Clocklab periodogram csv export file.}

\item{meta}{Path to the metadata file.}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
Periodogram table with metadata columns
}
\description{
Load Clocklab periodogram batch export table
}
\examples{
# TODO
}
