% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.boxPoint}
\alias{plot.boxPoint}
\title{Boxplot with data points}
\usage{
\method{plot}{boxPoint}(data, x, y, fill, theme = "classic")
}
\arguments{
\item{data}{Data}

\item{x}{x, must be already quoted}

\item{y}{y, must be already quoted}

\item{fill}{fill, must be already quoted}

\item{theme}{what theme to apply, see \code{plot.themes()}.}
}
\value{
ggplot object
}
\description{
Boxplot with data points
}
\examples{
# Internal use
}
