% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.linePoint}
\alias{plot.linePoint}
\title{Point plot with lines connecting 'mean' values.}
\usage{
\method{plot}{linePoint}(
  data,
  x,
  y,
  color = NA,
  theme = "classic",
  mean_func = mean,
  sd_func = function(x) stats::sd(x)/sqrt(length(x))
)
}
\arguments{
\item{data}{Data}

\item{x}{x, must be already quoted}

\item{y}{y, must be already quoted}

\item{color}{color, must be already quoted}

\item{theme}{what theme to apply, see \code{plot.themes()}}

\item{mean_func}{function to compute 'mean' (sample representative value)}

\item{sd_func}{function to compute 'sd' (sample spread)}
}
\value{
ggplot object
}
\description{
If \code{color} not specified or equal to \code{x}, whole data will be one color.
}
\examples{
# Internal use
}
