% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.themes}
\alias{plot.themes}
\title{Standardized themes for plotting}
\usage{
\method{plot}{themes}(type = "classic")
}
\arguments{
\item{type}{standardized theme name, see details.}
}
\value{
GG theme object.
}
\description{
Themes used by different plots in this package for standardized behavior.
}
\details{
Implemented themes:
\itemize{
\item \code{classic}: Based on \code{ggplot2::theme_classic()}.
}
}
\examples{
# Internal use
}
