% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.helpers.R
\name{plot_ggsave}
\alias{plot_ggsave}
\title{Simple wrapper of ggsave for saving plots.}
\usage{
plot_ggsave(
  plots = list(),
  widths,
  heights,
  units = "in",
  dir = "plots_eps",
  ...
)
}
\arguments{
\item{plots}{Named list of plots to save.}

\item{widths}{Width of the plots. Either length one or same as plots.}

\item{heights}{Height of the plots. Either length one or same as plots.}

\item{units}{Unit of width and height}

\item{dir}{Directory under which to save the plot}

\item{...}{These dots are for future extensions and must be empty.}
}
\description{
Simple wrapper of ggsave for saving plots.
}
\examples{
# TODO
}
