% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna-fish.R
\name{rfish_count}
\alias{rfish_count}
\title{Count number of dots per cell}
\usage{
rfish_count(samples, dots, ...)
}
\arguments{
\item{samples}{tibble of samples table, \code{rfish_read_samples()}}

\item{dots}{tibble of dots table, \code{rfish_read_dots()}}

\item{...}{<\code{\link[tidyselect:language]{tidy-select}}> extra columns (from samples table) to retain}
}
\value{
Count table (number of dots per cell).
}
\description{
Count number of dots per cell
}
\examples{
# TODO
}
