% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna-fish.R
\name{rfish_read_dots}
\alias{rfish_read_dots}
\title{Read dots table from snakemake pipeline output}
\usage{
rfish_read_dots(path, ...)
}
\arguments{
\item{path}{Path to the dots csv}

\item{...}{Forwarded to \code{readr::read_csv}}
}
\value{
tibble of dots table
}
\description{
Read dots table from snakemake pipeline output
}
\examples{
#TODO
}
