% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotInRoi-helpers.R
\name{spotInRoi_addMeta}
\alias{spotInRoi_addMeta}
\title{Add metadata column to spotInRoi data}
\usage{
spotInRoi_addMeta(df, idMap, meta, ...)
}
\arguments{
\item{df}{spotInRoi table, see details.}

\item{idMap}{identifier map table, see details.}

\item{meta}{metadata table.}

\item{...}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> columns to select from the metadata table.}
}
\value{
spotInRoi data with the specified metadata columns.
}
\description{
Add metadata column to spotInRoi data
}
\details{
spotInRoi data here takes a generalized meaning - a data frame with two
id columns $tp, $meas, plus any number of additional columns.

The identifier map table must contain two id columns $tp, $meas, plus
additional columns that are identifiers in the metadata table.
}
\examples{
# TODO
}
