% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnd.export.R
\name{spotInRoi_as.data.frame.df.nnd}
\alias{spotInRoi_as.data.frame.df.nnd}
\title{Expand NND results as data frame entries}
\usage{
spotInRoi_as.data.frame.df.nnd(df.nnd, nnd.column = "result")
}
\arguments{
\item{df.nnd}{Data frame with a NND result column}

\item{nnd.column}{Character name of the column that stores NND results.}
}
\value{
Expanded data frame with each NND value as one row
}
\description{
Expand NND results as data frame entries
}
\examples{
# TODO
}
