% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotInRoi-helpers.R
\name{spotInRoi_joinMeta}
\alias{spotInRoi_joinMeta}
\title{Join spotInRoi data with metadata columns}
\usage{
spotInRoi_joinMeta(df, df.meta, by, ...)
}
\arguments{
\item{df}{spotInRoi data frame.}

\item{df.meta}{metadata data frame.}

\item{by}{columns to perform join}

\item{...}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> columns to select from the metadata data frame.}
}
\value{
spotInRoi data frame joined with metadata.
}
\description{
Join spotInRoi data with metadata columns
}
\examples{
# Not exported
}
