% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnd.export.R
\name{spotInRoi_nnd.sort}
\alias{spotInRoi_nnd.sort}
\title{Sort spotInRoi data based on NND results}
\usage{
spotInRoi_nnd.sort(df.nnd, nnd.column = "result", ...)
}
\arguments{
\item{df.nnd}{spotInRoi data with NND results.}

\item{nnd.column}{Character name of the column that stores NND results.}

\item{...}{Passed to \code{base::order} for order permutation.}
}
\value{
Sorted spotInRoi results.
}
\description{
Sort spotInRoi data based on NND results
}
\examples{
# TODO
}
