% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{spotInRoi_read_csv}
\alias{spotInRoi_read_csv}
\title{Read data for spotInRoi analysis of the Puncta Tracker ImageJ plugin.}
\usage{
spotInRoi_read_csv(file, parse.spotOut = FALSE, ...)
}
\arguments{
\item{file}{File passed to \code{readr::read_csv}.}

\item{parse.spotOut}{Whether to parse spotOut array data.}

\item{...}{Other parameters passed to \code{readr::read_csv}.}
}
\value{
\code{tibble} of the spotInRoi analysis result.
}
\description{
Read data for spotInRoi analysis of the Puncta Tracker ImageJ plugin.
}
\examples{
# TODO
}
