% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotInRoi-helpers.R
\name{spotInRoi_selectNCT}
\alias{spotInRoi_selectNCT}
\title{Predefined subsetting of spotInRoi nuclear/full-cell data.}
\usage{
spotInRoi_selectNCT(dfNucCyt, what)
}
\arguments{
\item{dfNucCyt}{Parsed spotInRoi nuclear/full-cell data.}

\item{what}{What to subset. must be \code{nuclear}, \code{cytoplasm}, or \code{full}.}
}
\value{
Subsetted spotInRoi nuclear/full-cell data.
}
\description{
Predefined subsetting of spotInRoi nuclear/full-cell data.
}
\section{Subset defaults}{
Predefined subsetting rules are as follows:

\code{nuclear} - for each cell select the ROI with the smallest \code{equi.D}

\code{full} - for each cell select the ROI with the largest \code{equi.D}

\code{cytoplasm} - currently not implemented and will error.
}

\examples{
# TODO
}
