% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actogram.R
\name{subset_behavr}
\alias{subset_behavr}
\title{Simple subset of \code{behavr::behavr} behavior tables}
\usage{
subset_behavr(behavr, expr)
}
\arguments{
\item{behavr}{Behavior table to subset}

\item{expr}{An expression that evaluates to logical. See details.}
}
\value{
Subsetted behavior table.
}
\description{
Simple subset of \code{behavr::behavr} behavior tables
}
\details{
Right now the subsetting creates an intermediate logical vector,
which can be very large.
}
\examples{
# Not exported
}
