% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actogram.R
\name{actogram_loadFromMetadata}
\alias{actogram_loadFromMetadata}
\title{Read actogram metadata file}
\usage{
actogram_loadFromMetadata(path, ...)
}
\arguments{
\item{path}{Path to metadata, must be commonly used delimited or excel.}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
\code{behavr::behavr} behavior data table.
}
\description{
Read actogram metadata file
}
\examples{
# TODO
}
