% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctcf.R
\name{ctcf_plot}
\alias{ctcf_plot}
\title{Standardized plotting of CTCF data}
\usage{
ctcf_plot(data, x, y, fill = NA)
}
\arguments{
\item{data}{tibble of CTCF data}

\item{x}{x-axis variable for plotting}

\item{y}{y-axis variable for plotting}

\item{fill}{fill variable for plotting}
}
\value{
ggplot object
}
\description{
Standardized plotting of CTCF data
}
\examples{
# TODO
}
