% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.boxPoint}
\alias{plot.boxPoint}
\title{Boxplot with data points, viridis_d color filled}
\usage{
\method{plot}{boxPoint}(data, x, y, fill)
}
\arguments{
\item{data}{Data}

\item{x}{x, must be already quoted}

\item{y}{y, must be already quoted}

\item{fill}{fill, must be already quoted}
}
\value{
ggplot object
}
\description{
Boxplot with data points, viridis_d color filled
}
\examples{
# Internal use
}
